Remove GAlgebra Wrapper for
This commit simplifies the `GAlgebra` class of Droste. Instead of
being a final case class that wraps a `Function1` object, we change it
to just be an abstract class with an `apply` (or `run`) method.

Those using Scala 2.12 will benefit of  Single-Abstract-Method syntax,
to write most of the code without the `GAlgebra` constructor, like functions.