Remove GAlgebra Wrapper for
This commit simplifies the `GAlgebra` class of Droste. Instead of
being a final case class that wraps a `Function1` object, we change it
to just be an abstract class with an `apply` (or `run`) method.

Use of Scala 2.12 Single-Abstract-Method syntax allows us to write
most of the code without the `GAlgebra` constructor, like functions.